+++ src/App.tsx
@@ -1,6 +1,9 @@
 import { RouterProvider, createRouter } from "@tanstack/react-router"
 import "./App.css"
 
+import { ApiProvider, useApi } from "@/providers/api-provider"
+import { QueryClientProvider } from "@tanstack/react-query"
+import { queryClient } from "./lib/query-client"
 // 导入自动生成的路由树
 import { routeTree } from "./routeTree.gen"
 import { useAppStore } from "./stores/useAppStore"
@@ -12,6 +15,8 @@ const router = createRouter({
 	routeTree,
 	context: {
 		useStore: useAppStore, // 直接传递 hook 函数
+		queryClient, // 直接传递 queryClient 实例
+		useApi,
 	},
 })
 
@@ -23,7 +28,13 @@ declare module "@tanstack/react-router" {
 }
 
 function App() {
-	return <RouterProvider router={router} />
+	return (
+		<QueryClientProvider client={queryClient}>
+			<ApiProvider>
+				<RouterProvider router={router} />
+			</ApiProvider>
+		</QueryClientProvider>
+	)
 }
 
 export default App