+++ src/types/index.ts
@@ -1,5 +1,9 @@
+import type { ApiHooks } from "@/api/factory"
 import type { AppStoreType } from "@/stores/useAppStore"
+import type { QueryClient } from "@tanstack/react-query"
 
 export interface MyRouterContext {
 	useStore: AppStoreType
+	queryClient: QueryClient
+	useApi: () => ApiHooks
 }